requirements.txt
django
pipenv
djangorestframework
django-cors-headers
django-environ
spotipy